move abstract classes out of types
