+++ src/packages/tiptap/extensions/index.ts
@@ -1,1 +1,2 @@
+export * from './base.js';
 export type * from './types.js';
