+++ src/packages/ufm/filters/base.filter.ts
@@ -1,0 +1,6 @@
+import type { UmbUfmFilterApi } from '../ufm-filter.extension.js';
+
+export abstract class UmbUfmFilterBase implements UmbUfmFilterApi {
+	abstract filter(...args: Array<unknown>): string | undefined | null;
+	destroy() {}
+}
