+++ src/packages/ufm/index.ts
@@ -1,6 +1,7 @@
-export * from './types.js';
 export * from './components/index.js';
 export * from './contexts/index.js';
+export * from './filters/base.filter.js';
 export * from './plugins/index.js';
+export * from './types.js';
 export * from './ufm-component.extension.js';
 export * from './ufm-filter.extension.js';
